Corrected widths error.
I had forgotten about converting the 'face-to-face' width to a radius
before generating the cylinders. This change the code to take that into
account.

I also sloped the main body. Despite the mistake, the side of the viewer
that fit against the face was almost perfect, even if it wasn't what I
planned. Now, the front is better suited to match the phone and the back
should still fit against the face.