Move the 'top of slider' marks.
For some reason, the little cylinder on the outside slider caused a
glitch in slic3r that created a filled layer all the way across when
the little cylinder marker was on the lip. Moved to top without
side-effects. Made the same change to the inside slider, because it was
invisible on the original.