Add support for a lens description.
Lenses are a little too complicated to be described with a couple of
numbers to pass around. The file lenses.scad defines support functions
that allow us to treat a vector of numbers as if it we are descriptive
structure.

Define the three lenses we are currently working with.

Modify the code to use the new lens descriptor rather than just a lens
diameter.