Improve nose fit for lens plate.
The nose cut needs to be higher and more rounded on top. This is almost
as far as I'm comfortable and it's still not centered on the eyes.
The next stage would be to come up with a way to mount the lenses behind
the plate instead of in front of it.