Optimize build plates.
I'm trying to get as much as possible on a single build plate. It's not
working particularly well. The new lens holders won't print inside the
visor and the lens plates pretty much triple the area needed for the
print.
Even printing all of the optics on a single plate doesn't work well. I
may need to replace the single plate with 3 separate plates.