Move the strap mounts outward again.
As Jarrad requested, I'm moving the strap mounts out to the sides of the
visor. However, the original reason for moving them was that they kind
of looked "tacked on". So, I made the mount a bit longer and extended
the angle at the bottom. This makes it a little less blocky at the
bottom.

Hopefully that will help some.